refactor: Hrコンポーネントを装飾コンポーネントとして適切なディレクトリに移動し、Tailwind CSS優先方針を実装 fixes #22
