+++ src/app/components/common/decoration/Hr/Hr.tsx
@@ -1,15 +1,15 @@
 /**
- * @description Hrアイコンコンポーネント（水平線装飾）
+ * @description Hr装飾コンポーネント（水平線装飾）
  * @author WeddingInvitations
  * @since 1.0.0
  */
 
 import React from "react";
-import HrLeft from "../HrLeft/HrLeft";
-import HrRight from "../HrRight/HrRight";
+import HrLeft from "../../icon/HrLeft/HrLeft";
+import HrRight from "../../icon/HrRight/HrRight";
 
 /**
- * @description HrアイコンコンポーネントのProps型定義
+ * @description Hr装飾コンポーネントのProps型定義
  * @interface HrProps
  * @since 1.0.0
  */
@@ -26,15 +26,17 @@ interface HrProps {
   lineHeight?: "h-px" | "h-0.5" | "h-1" | "h-2" | "h-3" | "h-4";
   /** 水平線の色（Tailwindクラス） */
   lineColor?: string;
+  /** コンポーネントの高さ（Tailwindクラス） */
+  containerHeight?: "h-8" | "h-10" | "h-12" | "h-16" | "h-20" | "h-24" | "h-32";
 }
 
 /**
- * @description Hrアイコンコンポーネント（水平線装飾）
+ * @description Hr装飾コンポーネント（水平線装飾）
  * Figmaから取得したhrコンポーネントを再現
  * @param props - コンポーネントのProps
  * @returns JSX.Element
  * @example
- * <Hr iconSize={38} color="text-lavender-600" />
+ * <Hr iconSize={38} color="text-lavender-600" containerHeight="h-12" />
  */
 const Hr: React.FC<HrProps> = ({
   width,
@@ -43,10 +45,11 @@ const Hr: React.FC<HrProps> = ({
   className = "",
   lineHeight = "h-px",
   lineColor = "bg-lavender-600",
+  containerHeight = "h-12",
 }) => {
   return (
     <div
-      className={`relative flex items-center ${className}`}
+      className={`relative flex items-center ${containerHeight} ${className}`}
       style={width ? { width: `${width}px` } : undefined}
     >
       {/* 中央の水平線（背景） */}
