+++ src/app/components/common/decoration/Hr/index.ts
