+++ src/app/components/common/decoration/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @description 装飾コンポーネントのエクスポート
+ * @author WeddingInvitations
+ * @since 1.0.0
+ */
+
+export { default as Hr } from "./Hr/Hr";
